Add files via upload
In October 2015, Walt Hickey analyzed movie ratings data and found strong evidence to suggest that Fandango's rating system was biased and dishonest. He was able to prove this by comparing display results to ratings in the HTML of the web page. Fandango responded to the article saying it was due to a bug in the stystem and promised to fix the bug as soon as possible.

Objective of this project is to analyze the latest movie ratings data to determine if there has been any change since Hickey's analysis.